test(lang-v2): cover pod operator overflow behavior
